#494 dynamic model

+++ platform-core/src/main/java/com/webank/wecube/platform/core/jpa/PluginPackageAttributeRepository.java
@@ -20,6 +20,6 @@ public interface PluginPackageAttributeRepository extends CrudRepository<PluginP
     Optional<List<PluginPackageAttribute>> findAllReferenceByAttribute(
             @Param("packageName") String packageName,
             @Param("entityName") String entityName,
-            @Param("dataModelVersion") long dataModelVersion);
+            @Param("dataModelVersion") int dataModelVersion);
 
 }
